Backend_CSJ/requirements.txt
Django==4.1
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.13.1
drf-yasg==1.21.3
whitenoise==6.2.0
python-slugify==6.1.2
djangorestframework-simplejwt==5.2.0
django-filter==22.1
reportlab==3.6.11
trml2pdf==0.6
python-dateutil==2.8.2
Pillow==9.2.0
python-geoip==1.2
geoip2==4.6.0
python-geoip-geolite2==2015.303
gunicorn==20.1.0
django-cors-headers==3.13.0
python-ldap==3.4.2
pytest==7.2.0
